New send
+ APP-8: Address book in setting should be shown the same design address book in send screen
+ APP-1: Input field is shown undefine in send screen when clicking max button
