+++ src/screens/FrequentReceivers/index.js
@@ -2,8 +2,6 @@ import FrequentReceivers from './FrequentReceivers';
 
 export { default as FrequentReceiversForm } from './FrequentReceivers.form';
 
-export { default as FrequentReceiversAll } from './FrequentReceivers.all';
-
 export {
   default as withSyncIncognitoAddress,
 } from './FrequentReceivers.withSync';
